Add files via upload
FALTA HACER EL ULTIMO PUNTO DE LA ENTREGA ( EL 11) AHI EN PRIMERAENTREGATEST, YA ESTAN TODOS LOS ALGOMONES Y LOS ATAQUES, FRAN ESTABA HACIENDO LOS TEST DE ATAQUES, POR LO QUE NO SUBI MI VERSION (IGUALMENTE SON LOS QUE ESTABAN EN LO QUE SUBI AYER), SI SUBI TIPOTEST DE NUEVO PORQUE ESOS TEST FALTAN, SIGAN EL MODELO DEL EJEMPLO QUE HICE, Y ES NECESARIO BAJARSE ESTA ULTIMA VERSION POR LOS CAMBIOS DE LOS DOUBLE A ENTEROS, LO UNICO QUE FALTA ES LO QUE DIJE, EL PUNTO 11 DE LA ENTREGA, Y LOS TIPOTEST, PORQUE FRAN YA ESTABA HACIENDO LOS ATAQUES TEST

+++ Algomon.java
@@ -2,7 +2,7 @@ import java.util.HashMap;
 
 public class Algomon {
 	
-	double vida;
+	int vida;
 	HashMap<String, Ataque> ataques = new HashMap<String, Ataque>();
 	boolean vivo = true;
 	Tipo tipo;
@@ -11,8 +11,8 @@ public class Algomon {
 		return ataques.get(ataque).atacar(otroAlgomon);
 	}
 	
-	public void recibirAtaque(double daño){
-		double vidaRestante = vida - daño;
+	public void recibirAtaque(int daño){
+		int vidaRestante = vida - daño;
 		if (vidaRestante <= 0){
 			vida = 0;
 			vivo = false;
