+++ AtaqueSimple.java
@@ -1,11 +1,11 @@
 
 public class AtaqueSimple extends Ataque {
 	
-	double daño;
+	int daño;
 	
 	public boolean atacar(Algomon algomon){
 		if (cantidad > 0){
-			algomon.recibirAtaque(daño*(tipo.conTipo(algomon.getTipo())));
+			algomon.recibirAtaque((int) (daño*(tipo.conTipo(algomon.getTipo()))));
 			cantidad --;
 			return true;
 		}
